Add files via upload
A great module to learn about data cleaning.
Its a very important step just before performing data visualization. Its so important to have our data cleaned and in proper format so that we can visualize the data in the best possible manner. this module is for a beginner who is trying to learn how to clean the data.

#Dataset
The dataset that I have used is googleplaystoredata , its a data set of different apps of google play store and performing EDA on this dataset is a great learning opportunity for a beginner. I have download this data set from Kaggle and there are alot of different resources available online to help you explore more about this dataset.

#Libraries used
I have used the following libraries - 
Numpy
Seaborn
Matplotlib
Pandas

challenge in this dataset:
In this dataset the most challenge thing was to learn about the functions that help in cleaning the data and the kind of arguments to be passed into these functions.